feat: add `delete_existing` input

+++ src/index.ts
@@ -6,6 +6,7 @@ import {Client as NotionClient} from '@notionhq/client';
 import {appendNew, deleteExisting} from './sync';
 
 export interface Inputs {
+  delete_existing: boolean;
   file: string;
   notion_token: string;
   notion_id: string;
@@ -13,6 +14,7 @@ export interface Inputs {
 
 async function main() {
   const inputs: Inputs = {
+    delete_existing: core.getBooleanInput('delete_existing', {required: true}),
     file: core.getInput('file', {required: true}),
     notion_token: core.getInput('notion_token', {required: true}),
     notion_id: core.getInput('notion_id', {required: true}),
@@ -41,9 +43,11 @@ async function main() {
   core.info('Client created successfully.');
   core.endGroup();
 
-  core.startGroup('Deleting existing block');
-  await deleteExisting(client, inputs);
-  core.endGroup();
+  if (inputs.delete_existing) {
+    core.startGroup('Deleting existing block');
+    await deleteExisting(client, inputs);
+    core.endGroup();
+  } else core.info('Not trying to delete existing block.');
 
   core.startGroup('Appending new block');
   await appendNew(client, inputs, blocks);
